Moved source files to src/.
