+++ package.json
@@ -6,7 +6,7 @@
   "bin": "bin/index.js",
   "scripts": {
     "test": "NODE_ENV=test nyc mocha --recursive test/",
-    "lint": "eslint lib/",
+    "lint": "eslint lib/ test/",
     "clean": "rm -rf ./simian-complexity-data"
   },
   "repository": {
