+++ src/aggregate/calculate-aggregates.js
