+++ src/aggregate/index.js
