+++ src/aggregate/model.js
