+++ src/cli/command-spec.js
