+++ src/cli/print-usage-guide.js
