+++ src/utils/should-file-path-be-kept.js
